# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Agrocom API'
  description: 'API del Sistema de Gestión de Operaciones de Fumigación de Agrocom SRL. Sirve a las apps de campo (`agrocom-field`) y al portal del cliente; el spec generado desde el código es el contrato de la API (ADR 0014). Dinero y hectáreas viajan como string decimal, nunca como número de punto flotante.'
  version: 0.1.0
servers:
  -
    url: /
    description: 'Servidor actual — la URL base la define el entorno (local, staging, producción).'
paths:
...
  /api/sync/catalogo:
    get:
      tags:
        - Sincronizacion
      summary: 'Pull de catálogo con cursor (órdenes, lotes, personas, trabajos)'
      description: 'Baja el catálogo de órdenes vigentes, lotes, personas y trabajos asignados desde el panel modificados desde la posición del cursor recibido, con paginación por cursor (`updated_at`, `id`) — nunca por número de página, para no perder ni repetir registros entre pulls. `desde` vacío o ausente trae todo lo vigente (primera sincronización). Un `desde` no decodificable se trata igual que vacío, nunca como error de validación.'
      operationId: obtenerCatalogoSincronizacion
      parameters:
        -
          name: desde
          in: query
          description: 'Cursor opaco devuelto por el pull anterior. Ausente o vacío en la primera sincronización.'
          required: false
          schema:
            type: string
      responses:
        200:
          description: 'Catálogo modificado desde el cursor, y el cursor de continuación para el próximo pull.'
          content:
            application/json:
              schema:
                required:
                  - ordenes
                  - lotes
                  - personas
                  - trabajos
                  - cursor
                properties:
                  ordenes:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrdenCatalogo'
                  lotes:
                    type: array
                    items:
                      $ref: '#/components/schemas/LoteCatalogo'
                  personas:
                    type: array
                    items:
                      $ref: '#/components/schemas/PersonaCatalogo'
                  trabajos:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrabajoCatalogo'
                  cursor:
                    description: 'Cursor opaco a reenviar tal cual en `desde` en el próximo pull.'
                    type: string
                type: object
        401:
          description: 'Token ausente, revocado o ya sin rol válido.'
      security:
        -
          tokenDispositivo: []
  /api/sync:
    post:
      tags:
        - Sincronizacion
      summary: 'Push de sincronización en lote (trabajo, sesión)'
      description: 'Push en lote de la cola offline (espec §2.1). Cada registro se procesa en su propia transacción —nunca el lote completo en una— y se responde con uno de tres estados: `aplicado`, `duplicado` (reintento del mismo `uuid_cliente` ya aplicado, se trata como éxito) o `rechazado` con motivo. El servidor agrupa por tipo y aplica siempre en el orden `trabajo`, `sesion`, `cierre_trabajo`, `cierre_sesion`, sin importar el orden del arreglo recibido, así que un registro puede referenciar otro del mismo lote aunque venga antes en el arreglo.'
      operationId: sincronizarLote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - registros
              properties:
                registros:
                  type: array
                  items:
                    $ref: '#/components/schemas/RegistroSync'
              type: object
      responses:
        200:
          description: 'Resultado por registro, en el mismo orden del arreglo de entrada.'
          content:
            application/json:
              schema:
                required:
                  - resultados
                properties:
                  resultados:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResultadoSync'
                type: object
        401:
          description: 'Token ausente, revocado o ya sin rol válido.'
        422:
          description: 'Falta `registros`, o no es un arreglo.'
      security:
        -
          tokenDispositivo: []
components:
  schemas:
...
    LoteDeOrdenCatalogo:
      title: 'Lote de una orden (catálogo)'
      description: 'Uno de los lotes que cubre una orden, con las hectáreas que le pidió a ESE lote (HU-92, tarea 107 — una orden puede cubrir varios lotes de la propiedad). DECIMAL como string (invariante 6).'
      required:
        - lote_id
        - hectareas_solicitadas
      properties:
        lote_id:
          description: 'Lote que cubre la orden (módulo Comercial, solo ID).'
          type: integer
          example: 3
        hectareas_solicitadas:
          description: 'Hectáreas que la orden le pidió a ESE lote. DECIMAL como string.'
          type: string
          example: '50.00'
      type: object
    OrdenCatalogo:
      title: 'Orden de aplicación (catálogo)'
      description: 'Orden vigente para el pull de catálogo (espec §4.3, ampliada HU-92 tarea 107: `lotes` reemplaza el `lote_id` único de antes; HU-79 tarea 110: `litros_ha`/`kilos_por_vuelo` son mutuamente excluyentes según la categoría de insumo). Los DECIMAL viajan como string (invariante 6).'
      required:
        - id
        - contrato_id
        - lotes
        - nro_aplicacion
        - litros_ha
        - kilos_por_vuelo
        - humedad_min_pct
        - viento_max_kmh
        - temperatura_max_c
        - humedad_max_pct
        - velocidad_max_kmh
        - altura_vuelo_m
        - velocidad_vuelo_kmh
        - ancho_pasada_m
        - observaciones
        - emitida_por_contacto_id
        - fecha_emision
        - estado
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        contrato_id:
          type: integer
          example: 1
        lotes:
          type: array
          items:
            $ref: '#/components/schemas/LoteDeOrdenCatalogo'
        nro_aplicacion:
          type: integer
          example: 1
        litros_ha:
          description: 'Dosis en litros por hectárea (insumo líquido); null si la categoría es sólida.'
          type: string
          example: '10.00'
          nullable: true
        kilos_por_vuelo:
          description: 'Dosis en kilos por vuelo (insumo sólido); null si la categoría es líquida.'
          type: string
          example: '8.50'
          nullable: true
        humedad_min_pct:
          type: string
          example: '60.00'
          nullable: true
        viento_max_kmh:
          type: string
          example: '15.00'
          nullable: true
        temperatura_max_c:
          type: string
          example: '32.00'
          nullable: true
        humedad_max_pct:
          type: string
          example: '90.00'
          nullable: true
        velocidad_max_kmh:
          type: string
          example: '25.00'
          nullable: true
        altura_vuelo_m:
          type: string
          example: '3.00'
          nullable: true
        velocidad_vuelo_kmh:
          type: string
          example: '18.00'
          nullable: true
        ancho_pasada_m:
          type: string
          example: '7.00'
          nullable: true
        observaciones:
          type: string
          example: 'Aplicar en horas de la mañana.'
          nullable: true
        emitida_por_contacto_id:
          type: integer
          example: 2
          nullable: true
        fecha_emision:
          type: string
          format: date
          example: '2026-08-26'
        estado:
          type: string
          example: vigente
        updated_at:
          type: string
          format: date-time
          example: '2026-08-26T12:00:00+00:00'
      type: object
    LoteCatalogo:
      title: 'Lote (catálogo)'
      description: 'Lote vigente (no borrado) para el pull de catálogo (espec §4.1). `hectareas` es DECIMAL como string (invariante 6).'
      required:
        - id
        - campo_id
        - codigo
        - hectareas
        - geometria
        - restricciones
        - updated_at
      properties:
        id:
          type: integer
          example: 3
        campo_id:
          type: integer
          example: 1
        codigo:
          type: string
          example: L-01
        hectareas:
          type: string
          example: '120.50'
        geometria:
          description: 'GeoJSON del lote, o null si no está cargado.'
          type: object
          nullable: true
        restricciones:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
          example: '2026-08-26T12:00:00+00:00'
      type: object
    PersonaCatalogo:
      title: 'Persona operativa (catálogo)'
      description: 'Persona vigente (no borrada) para el pull de catálogo (espec §4.2).'
      required:
        - id
        - nombre
        - rol
        - base_id
        - activo
        - updated_at
      properties:
        id:
          type: integer
          example: 5
        nombre:
          type: string
          example: 'Piloto Uno'
        rol:
          type: string
          example: piloto
        base_id:
          type: integer
          example: 1
          nullable: true
        activo:
          type: boolean
          example: true
        updated_at:
          type: string
          format: date-time
          example: '2026-08-26T12:00:00+00:00'
      type: object
    TrabajoCatalogo:
      title: 'Trabajo asignado desde el panel (catálogo)'
      description: 'Trabajo abierto por el jefe de campo al repartir una orden vigente entre equipos (HU-70, tarea 85) — nunca los que nacen por sync (`equipo_trabajo_id` siempre presente acá). `uuid_cliente` es el que generó el panel al confirmar la asignación: la app lo usa TAL CUAL para abrir sesiones sobre este trabajo. `hectareas_declaradas` es DECIMAL como string (invariante 6).'
      required:
        - id
        - uuid_cliente
        - orden_id
        - lote_id
        - hectareas_declaradas
        - equipo_trabajo_id
        - updated_at
      properties:
        id:
          type: integer
          example: 42
        uuid_cliente:
          type: string
          example: 9a1b7e3e-2f7a-4b3d-8c1e-6f2a1d9c4b0a
        orden_id:
          type: integer
          example: 1
        lote_id:
          type: integer
          example: 3
        hectareas_declaradas:
          type: string
          example: '300.00'
        equipo_trabajo_id:
          type: integer
          example: 7
        updated_at:
          type: string
          format: date-time
          example: '2026-09-13T12:00:00+00:00'
      type: object
...
    ResultadoSync:
      title: 'Resultado de aplicar un registro del lote'
      description: 'Uno por registro de entrada, en el mismo orden (espec §2.1, punto 3).'
      required:
        - uuid_cliente
        - tipo
        - estado
      properties:
        uuid_cliente:
          type: string
          example: a1b2c3d4-0000-4000-8000-000000000001
          nullable: true
        tipo:
          type: string
          example: trabajo
          nullable: true
        estado:
          type: string
          example: aplicado
          enum:
            - aplicado
            - duplicado
            - rechazado
        motivo:
          description: 'Presente solo cuando `estado` es `rechazado`.'
          type: string
          example: 'el trabajo referenciado no existe todavía'
      type: object
  securitySchemes:
    tokenDispositivo:
      type: http
      description: 'Token por dispositivo emitido por `POST /api/auth/token` (HU-03). Se envía como `Authorization: Bearer {token}`. No caduca por tiempo: vale hasta que se lo revoque desde el panel, o hasta que la cuenta o el rol con el que se emitió dejen de estar vigentes.'
      scheme: bearer
...